% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBaseNamespace.R
\docType{data}
\name{base-environments}
\alias{base-environments}
\alias{isBaseNamespace}
\alias{baseenv}
\alias{.BaseNamespaceEnv}
\title{Base environments}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
isBaseNamespace(ns)

baseenv()

.BaseNamespaceEnv
}
\arguments{
\item{ns}{string or namespace environment.}
}
\value{
A boolean
}
\description{
\itemize{
\item \code{.BaseNamespaceEnv} is the namespace environment for the base package, it is
a child of the global environment and an indirect parent to all other
naturally defined namespaces.
\item The  \code{baseenv()} function is a primitive that returns the environment of the base package,
an ancestor of the global environment). It contains rigorously the same elements
as \code{.BaseNamespaceEnv} and modifying one applies the same modifications to the
other (a behavior that doesn't affect other packages)
\item \code{isBaseNamespace()} checks if the input is identical to \code{.BaseNamespaceEnv}.
}
}
\details{
{rlang} has \code{rlang::base_env()} which is completely identical to \code{base.env()},
but respecting tidyverse naming conventions.
}
\keyword{datasets}
