% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knownS3Generics.R
\name{.knownS3Generics}
\alias{.knownS3Generics}
\title{S3 generics defined in base R}
\description{
This a named vector where names are the names of the generics defined in base R
and the value is the bas package they're defined on. We find generics defined
in {base}, {utils}, {graphics} and {stats}.
}
\examples{
split(names(.knownS3Generics), .knownS3Generics)
}
