% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeImportMethods.R
\name{.mergeImportMethods}
\alias{.mergeImportMethods}
\title{.mergeImportMethods}
\usage{
.mergeImportMethods(impenv, expenv, metaname)
}
\arguments{
\item{impenv}{environment.}

\item{expenv}{namespace environment.}

\item{metaname}{the methods table name.}
}
\description{
WIP
}
