% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamespace.R
\name{getNamespace}
\alias{getNamespace}
\alias{asNamespace}
\alias{.getNamespace}
\alias{..getNamespac}
\title{Get a namespace from its name}
\usage{
getNamespace(name)

asNamespace(ns, base.OK = TRUE)

.getNamespace(name)

..getNamespac(name, where)
}
\arguments{
\item{name}{A string or a symbol}

\item{ns}{A string, a symbol or a namespace}

\item{base.OK}{A boolean, whether to fail if \code{ns} refers to the base namespace}

\item{where}{A string naming the object mentioned in the warning, can be set to "unknown" to suppress the warning (unless
the environment variable "\emph{R_NO_REPORT_MISSING_NAMESPACES}" was set),}
}
\description{
These are 4 very similar functions. Note: we needed to mispell \code{..getNamespace}
to get around a \code{devtools::document()} bug
}
\details{
\itemize{
\item \code{.getNamespace()} is the lowest level version, it returns the environment representing the name space name.
It returns \code{NULL} if the namespace wasn't not loaded
\item \code{getNamespace()} returns the environment representing the name space name.
The namespace is loaded if necessary and the function crashes if it can't be loaded.
\item \code{..getNamespace()} is similar but warns instead of failing and falls back on \code{.GlobalEnv},
the warning can be turned of by setting the environment variable "\emph{R_NO_REPORT_MISSING_NAMESPACES}"
to "false" or setting \code{where} to "unknown" if the environment variable wasn't set.
\item \code{asNamespace()} is a wrapper around \code{getNamespace()} that can be provided a
namespace directly (it is then returned as is) and has an argument \code{base.OK}
to fail if \code{ns} refers to the base namespace.
}
}
\examples{
nsName <- "stats"
(ns <- asNamespace(nsName))
}
\seealso{
the inverse function \code{environmentName()}
}
