% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.object.R
\name{is.object}
\alias{is.object}
\title{Does the object have a class attribute ?}
\usage{
is.object(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
A primitive function that tests efficiently the existence of a class attribute.
Note that \code{class()} doesn't necessarily return the class attribute but might
return the so called "implicit class". Use \code{oldClass()} or \code{attr(x, "class")}
to fetch the class attribute.
}
\details{
The term \code{object} is a bit unfortunate here as it is widely used in the official
doc regardless of class attribute, e.g. in \code{?length} \code{x} is described as
"an R object" and the title of the canonical help page for \code{is.object}
is an oxymoron. It might be clearer to talk about classless and classed objects.
}
