% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingNamespaceInfo.R
\name{loadingNamespaceInfo}
\alias{loadingNamespaceInfo}
\title{loadingNamespaceInfo}
\usage{
loadingNamespaceInfo()
}
\description{
Checks the existence of the \verb{__LoadingNamespaceInfo__} object in the callers and fails if not found
}
\examples{
try(loadingNamespaceInfo())
fun <- function() {
  `__LoadingNamespaceInfo__` <- "test"
  loadingNamespaceInfo()
}
}
