% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaceExport.R
\name{namespaceExport}
\alias{namespaceExport}
\title{Export objects in a namescape}
\usage{
namespaceExport(ns, vars)
}
\arguments{
\item{ns}{An unlocked namespace}

\item{vars}{A character vector}
}
\description{
For a given unlocked namespace \code{ns}, export the previously unexported objects named as \code{vars}.
Note that namespace are almost always locked and that there is no way to unlock them with base R.
The only moment a namespace is unlocked is when the \code{.onLoad()} function is called. However
the package {rlang} provides a function \code{env::unlock()} that allows unlocking environments.
}
\examples{
\dontrun{
# build toy namespace
ns <- asNamespace("utils")
rlang::env_unlock(ns)
"head.array" \%in\% getNamespaceExports(ns)
base::namespaceExport(ns, "head.array")
"head.array" \%in\% getNamespaceExports(ns)
}
}
