% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{new_doc}
\alias{new_doc}
\alias{sync}
\alias{pr}
\title{Open or create a new documentation file}
\usage{
new_doc(f, alt.doc.path = getOption("alt.doc.path"))

sync(alt.doc.path = getOption("alt.doc.path"))

pr(
  branch,
  message = sprintf("document \%s", branch),
  alt.doc.path = getOption("alt.doc.path")
)
}
\arguments{
\item{f}{function name, unquoted}

\item{alt.doc.path}{path to {alt.doc}'s package folder}

\item{message}{commit message}
}
\description{
These functions are meant to be called from any environment, not necessarily
while actively developing {alt.doc}.
}
\details{
\itemize{
\item \code{alt.doc::new_doc(f)} creates an R script in the {alt.doc} package.
\item \code{alt.doc::sync()} rebuilds the doc and reinstalls the package silently.
\item \code{alt.doc::pr()} commit changes to a newly created branch and opens a page to PR to the repo,
thens checks out the main branch back.
}

The environment of this function is set to .BaseNamespaceEnv to avoid
conflicts
}
