% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHasNamespace.R
\name{packageHasNamespace}
\alias{packageHasNamespace}
\title{Check if a package has a NAMESPACE file}
\usage{
packageHasNamespace(package, package.lib)
}
\arguments{
\item{package}{string naming the package/namespace to load.}

\item{package.lib}{character vector specifying library, such as provided by \code{ .libPaths()}}
}
\value{
a logical vector as long as package.lib
}
\description{
\code{packageHasNamespace()} does not indicate if the package has a namespace
(all now do), rather if it has a NAMESPACE file, which base and some legacy
packages do not.
}
\examples{
packageHasNamespace("alt.doc", .libPaths())
packageHasNamespace("stats", .libPaths())
packageHasNamespace("base", .libPaths())
packageHasNamespace("not a package name", .libPaths())
}
