% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseNamespaceFile.R
\name{parseNamespaceFile}
\alias{parseNamespaceFile}
\title{Parse the NAMESPACE file}
\usage{
parseNamespaceFile(package, package.lib, mustExist = TRUE)
}
\arguments{
\item{package}{the name of a package}

\item{package.lib}{path to the library, such as an element of \code{.libPaths()} or
a folder containing packages under development}
}
\description{
Parse a NAMESPACE file into a named list containing :
}
\details{
\itemize{
\item imports: A list of character vectors, length 1 for import(PKG) and length 2 for importFrom(PKG, FUN)
\item exports: A character vector of exported object names
\item exportPatterns: A character vector
\item importClasses: A list
\item importMethods: A list
\item exportClasses: A character vector
\item exportMethods: A character vector
\item exportClassPatterns: A character vector
\item dynlibs: A string
\item nativeRoutines: A list
\item S3methods: A matrix
}
}
\examples{
parseNamespaceFile("stats", .libPaths()[[2]])
}
