% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerS3method.R
\name{registerS3method}
\alias{registerS3method}
\title{Register S3 Methods}
\usage{
registerS3method(genname, class, method, envir = parent.frame())
}
\arguments{
\item{genname}{a character string naming an S3 generic function.}

\item{class}{a character string naming an S3 class.}

\item{method}{a character string or function giving the S3 method to be registered.}

\item{envir}{The environment where to fetch \code{method} from if it's a string (ignored otherwise)}
}
\description{
Very similar to \code{.S3method()}, which is properly documented and close wrapper
around \code{registerS3method()}. The difference is that \code{method} is optional for \code{.S3method()}
and \code{envir} can only be set in \code{registerS3method()}.
}
