% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamespaceInfo.R
\name{setNamespaceInfo}
\alias{setNamespaceInfo}
\title{Set the namespace information}
\usage{
setNamespaceInfo(ns, which, val)
}
\arguments{
\item{ns}{string or namespace environment.}

\item{which}{The name of the namespace information item to set.}

\item{val}{The value of the set item.}
}
\description{
This defines or overrides items in the \code{.__NAMESPACE__.} object of a namespace.
}
\seealso{
getNamsepaceInfo
}
