\name{test}
\alias{test}
\title{test}

\section{MySection1}{

Technically any \code{getNamespaceInfo()} could access any item define in \code{ns[[".__NAMESPACE__."]]}
(e.g. through \code{setNamespaceInfo()}) but in practice the namespace info is made
from the following components
}

\section{MySection2}{

Technically any \code{getNamespaceInfo()} could access any item define in \code{ns[[".__NAMESPACE__."]]}
(e.g. through \code{setNamespaceInfo()}) but in practice the namespace info is made
from the following components
}
